+++ autofit/__init__.py
@@ -30,10 +30,10 @@ from .mapper.model_mapper import ModelMapper
 from .mapper.model_mapper import ModelMapper as Mapper
 from .mapper.model_object import ModelObject
 from .mapper.operator import DiagonalMatrix
-from .mapper.prior.assertion import ComparisonAssertion
-from .mapper.prior.assertion import ComparisonAssertion
-from .mapper.prior.assertion import GreaterThanLessThanAssertion
-from .mapper.prior.assertion import GreaterThanLessThanEqualAssertion
+from .mapper.prior.arithmetic.assertion import ComparisonAssertion
+from .mapper.prior.arithmetic.assertion import ComparisonAssertion
+from .mapper.prior.arithmetic.assertion import GreaterThanLessThanAssertion
+from .mapper.prior.arithmetic.assertion import GreaterThanLessThanEqualAssertion
 from .mapper.prior.deferred import DeferredArgument
 from .mapper.prior.deferred import DeferredInstance
 from .mapper.prior.width_modifier import AbsoluteWidthModifier
@@ -84,12 +84,12 @@ from .text import samples_text
 from .tools import util
 
 
-from autofit.mapper.prior.compound import SumPrior as Add
-from autofit.mapper.prior.compound import MultiplePrior as Multiply
-from autofit.mapper.prior.compound import DivisionPrior as Divide
-from autofit.mapper.prior.compound import ModPrior as Mod
-from autofit.mapper.prior.compound import PowerPrior as Power
-from autofit.mapper.prior.compound import AbsolutePrior as Abs
+from autofit.mapper.prior.arithmetic.compound import SumPrior as Add
+from autofit.mapper.prior.arithmetic.compound import MultiplePrior as Multiply
+from autofit.mapper.prior.arithmetic.compound import DivisionPrior as Divide
+from autofit.mapper.prior.arithmetic.compound import ModPrior as Mod
+from autofit.mapper.prior.arithmetic.compound import PowerPrior as Power
+from autofit.mapper.prior.arithmetic.compound import AbsolutePrior as Abs
 
 from . import example as ex
 from . import database as db
