+++ autofit/mapper/prior/arithmetic/__init__.py
@@ -1,0 +1,1 @@
+from .arithmetic import ArithmeticMixin
