+++ autofit/mapper/prior/arithmetic/compound.py
